requirements.txt
Flask>=2.0
paho-mqtt>=1.6

